+++ tests/files/c_source/padding_bit_order.h
@@ -33,7 +33,6 @@
 
 #include <stdint.h>
 #include <stdbool.h>
-#include <unistd.h>
 
 #ifndef EINVAL
 #    define EINVAL 22
@@ -180,7 +179,7 @@ struct padding_bit_order_msg4_t {
  *
  * @return Size of packed data, or negative error code.
  */
-ssize_t padding_bit_order_msg0_pack(
+int padding_bit_order_msg0_pack(
     uint8_t *dst_p,
     const struct padding_bit_order_msg0_t *src_p,
     size_t size);
@@ -316,7 +315,7 @@ bool padding_bit_order_msg0_c_is_in_range(uint16_t value);
  *
  * @return Size of packed data, or negative error code.
  */
-ssize_t padding_bit_order_msg1_pack(
+int padding_bit_order_msg1_pack(
     uint8_t *dst_p,
     const struct padding_bit_order_msg1_t *src_p,
     size_t size);
@@ -452,7 +451,7 @@ bool padding_bit_order_msg1_h_is_in_range(uint16_t value);
  *
  * @return Size of packed data, or negative error code.
  */
-ssize_t padding_bit_order_msg2_pack(
+int padding_bit_order_msg2_pack(
     uint8_t *dst_p,
     const struct padding_bit_order_msg2_t *src_p,
     size_t size);
@@ -561,7 +560,7 @@ bool padding_bit_order_msg2_k_is_in_range(uint8_t value);
  *
  * @return Size of packed data, or negative error code.
  */
-ssize_t padding_bit_order_msg3_pack(
+int padding_bit_order_msg3_pack(
     uint8_t *dst_p,
     const struct padding_bit_order_msg3_t *src_p,
     size_t size);
@@ -616,7 +615,7 @@ bool padding_bit_order_msg3_l_is_in_range(uint64_t value);
  *
  * @return Size of packed data, or negative error code.
  */
-ssize_t padding_bit_order_msg4_pack(
+int padding_bit_order_msg4_pack(
     uint8_t *dst_p,
     const struct padding_bit_order_msg4_t *src_p,
     size_t size);
